Modify Block to use BlockType object
Block previously store the name of its blocktype, to make serialization
simple. However, this makes it necessary to often look up the block type
in many places. Now, Block stores a BlockType object, and the
serialization code has been moved into a class that must be instantiated
with the blocktypes so it can do the conversions.